+++ NEWS.md
@@ -1,5 +1,6 @@
 # ellmer (development version)
 
+* [BREAKING CHANGE] `type_array()` and `type_enum()` now have the description as the second argument and `items/`/`values` as the first. This makes them easier to use in the common case where the description isn't necessary (#610).
 * By default, ellmer now retries requests up to 3 times, controllable with `option(ellmer_max_tries)` and will retry if the connection fails (rather than just if the request itself returns a transient error).
 * The default timeout, controlled by `option(ellmer_timeout_s)`, now applies to the initial connection phase. 
 * [BREAKING CHANGE] `tool()` has a simpler specification: you now specify the `name`, `description`, and `arguments`. I have done my best to deprecate old usage and give clear errors, but I have likely missed a few edge cases. I apologise for the pain this causes, but I'm convinced that it is going to making tool usage easier and clearer. If you have many calls to convert, `?tool` contains a prompt that will help you use an LLM to convert them. (#603).